requirements.txt
Django==2.2.4
django-taggit==1.1.0
django-ckeditor==5.7.1
django-imagekit==4.0.2